fix(大屏-班组画像): 班组总的检修时长修改和个人的检修时长修改

+++ aiurt-module-inspection/aiurt-module-inspection-api/src/main/java/com/aiurt/boot/index/dto/TaskUserDTO.java
@@ -11,5 +11,5 @@ public class TaskUserDTO {
     private String taskId;
 
     @ApiModelProperty("检修工时")
-    private BigDecimal inspecitonTotalTime;
+    private Integer inspecitonTotalTime;
 }
