+++ aiurt-module-inspection/aiurt-module-inspection-api/src/main/java/com/aiurt/boot/index/dto/TeamPortraitDTO.java
@@ -32,7 +32,7 @@ public class TeamPortraitDTO {
     @ApiModelProperty("班组平均维修响应时间")
     private Integer averageTime;
     @ApiModelProperty("检修总工时")
-    private BigDecimal inspecitonTotalTime;
+    private Integer inspecitonTotalTime;
     /**巡检总工时，单位秒*/
     @ApiModelProperty("巡检总工时")
     private Integer patrolTotalTime;
